worker-k8s/dispatcher/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
black==22.12.0
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==39.0.0
firebase-admin==6.0.1
Flask==2.2.2
frozenlist==1.3.3
google-api-core==2.11.0
google-api-python-client==2.71.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-cloud-bigquery==3.4.1
google-cloud-core==2.3.2
google-cloud-firestore==2.7.3
google-cloud-pubsub==2.13.11
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.1
grpc-google-iam-v1==0.12.4
grpcio==1.51.1
grpcio-status==1.51.1
httplib2==0.21.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kubernetes==25.3.0
MarkupSafe==2.1.1
msgpack==1.0.4
multidict==6.0.4
mypy-extensions==0.4.3
oauthlib==3.2.2
packaging==21.3
pathspec==0.10.3
platformdirs==2.6.2
proto-plus==1.22.1
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.13
websocket-client==1.4.2
Werkzeug==2.2.2
yarl==1.8.2
